+++ tests/test_blocks/test_low_rank_block.py
