+++ tests/test_loss/test_power_loss.py
