+++ tests/test_model/test_average_neural_operator.py
